requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
playwright
apscheduler
alembic
passlib[bcrypt]
python-multipart